Corrections mineures

+++ Projet_C_Diff/Header.h
@@ -6,7 +6,7 @@ int strcomp1(const char* s1, const char* s2);
 FILE* OpenAFile(char* nomFichier, int optionOpen);
 unsigned long linesOfFile(char* file);
 unsigned long linesOfTab(char *tab, unsigned long* length);
-int CaractersOfFile(char* file);
+unsigned long CaractersOfFile(char* file);
 char* fileToTabs(char* file, int optionSpe, int optionN);
 char* fileToTabsOptionT(char* file, int optionSpe);
 char* pointFile(FILE* file);
